Improve the fix for xterm movement key behavior.
Add s:needspecialkeyhandling boolean and an extra map for xterm key
codes (s:movementKeysXterm), which need the same mapping. They get
mapped to the "proper" names.
I am not sure if this should not just go to somebodys vimrc, but without
autoclose the keys work properly.

Also add PageUp and PageDown to the list of movement keys.